Add an allowed version parameter for cf_release inlusion
[#143452901]

+++ tasks/upload-bosh-blobs-to-cf-release/run.rb
@@ -16,6 +16,7 @@ def buildpack_destination_dir(buildpack)
 end
 
 buildpack = ENV.fetch('BUILDPACK')
+version_allowed = ENV.fetch('VERSION_ALLOWED', '')
 version = ''
 
 blob_store_private_yml = {
@@ -38,26 +39,30 @@ Dir.chdir('cf-release') do
   version = matches[1] if matches.size > 1
   puts "Version for #{buildpack} is #{version}"
 
-  destination_dir = buildpack_destination_dir(buildpack)
-  system "rm -f blobs/#{destination_dir}"
-  blobs = YAML.load(File.read('config/blobs.yml'))
+  if version_allowed == '' || Gem::Dependency.new('', version_allowed).match?('', version)
+    destination_dir = buildpack_destination_dir(buildpack)
+    system "rm -f blobs/#{destination_dir}"
+    blobs = YAML.load(File.read('config/blobs.yml'))
 
-  old_buildpack_key = find_buildpack_key blobs, buildpack
+    old_buildpack_key = find_buildpack_key blobs, buildpack
 
-  next unless old_buildpack_key
-  new_sha = Digest::SHA1.file(buildpack_blob).hexdigest
+    next unless old_buildpack_key
+    new_sha = Digest::SHA1.file(buildpack_blob).hexdigest
 
-  next unless new_sha != blobs[old_buildpack_key]['sha']
-  blobs.delete(old_buildpack_key)
-  File.write('config/blobs.yml', YAML.dump(blobs))
-  exit 1 unless system "bosh add blob #{buildpack_blob} #{destination_dir}"
+    next unless new_sha != blobs[old_buildpack_key]['sha']
+    blobs.delete(old_buildpack_key)
+    File.write('config/blobs.yml', YAML.dump(blobs))
+    exit 1 unless system "bosh add blob #{buildpack_blob} #{destination_dir}"
 
-  exit 1 unless system "bosh -n upload blobs"
-  exit 1 unless system "/usr/bin/env bash ./scripts/setup-git-hooks"
+    exit 1 unless system "bosh -n upload blobs"
+    exit 1 unless system "/usr/bin/env bash ./scripts/setup-git-hooks"
 
-  GitClient.update_submodule_to_latest(buildpack_bosh_dir, cf_release_buildpack_submodule_dir)
-  GitClient.add_everything
-  GitClient.safe_commit("Update #{buildpack}-buildpack to v#{version}")
+    GitClient.update_submodule_to_latest(buildpack_bosh_dir, cf_release_buildpack_submodule_dir)
+    GitClient.add_everything
+    GitClient.safe_commit("Update #{buildpack}-buildpack to v#{version}")
+  else
+    puts "Skipping since not version #{version_allowed}"
+  end
 end
